Added Dt returned tensor change to type float32 for consistency.
Former-commit-id: 35afb483c9b64cf3449bf4dbc126e5641223a515

+++ src/yoke/datasets/lsc_dataset.py
@@ -710,7 +710,8 @@ class LSC_rho2rho_temporal_DataSet(Dataset):
 
         # Get the time offset
         Dt = 0.25 * (endIDX - startIDX)
-
+        Dt = Dt.to(torch.float32)
+        
         # Close the npzs
         start_npz.close()
         end_npz.close()
